Fixed Brand scrape
Removes 'Visit the ______ Store' so that it logs just the brand name, not the whole phrase. Also added a tryCount variable that only lets the script request the same URL 5 times before giving up and moving on to the next time - lets the program finish even if it encounters a broken link that it can't scrape the title from.